CLExpressionLike extends Immutable
Since Immutable has an unimplemented copy() method, this means a lot of classes have to implement this method.
The purpose of this at this point is simply for testing  the high-level laws that specify the behavior of equals overrides.
This is not a very good reason for requiring this method to be implemented in all this immutable classes, since there
is no actual use for the copy() method. It would be better to have another way to test equals().